Revert "Merge pull request #1 from TheD1rkMtr/main"
This reverts commit 46efae96a30783458f17a43c6bbaabc4673517d2, reversing
changes made to 51958e655adf2dbce8bfc356d1de77ca69c88c94.
