+++ .gitignore
@@ -1,0 +1,5 @@
+﻿build
+*.sln
+*.vcxproj
+*.filters
+*.user
