--- PatchingAPI/PatchingAPI.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32106.194
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PatchingAPI", "PatchingAPI.vcxproj", "{81E60DC6-694E-4F51-88FA-6F481B9A4208}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Debug|x64.ActiveCfg = Debug|x64
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Debug|x64.Build.0 = Debug|x64
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Debug|x86.ActiveCfg = Debug|Win32
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Debug|x86.Build.0 = Debug|Win32
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Release|x64.ActiveCfg = Release|x64
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Release|x64.Build.0 = Release|x64
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Release|x86.ActiveCfg = Release|Win32
-		{81E60DC6-694E-4F51-88FA-6F481B9A4208}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C4C56CBC-05A3-4229-8BCC-D84AB88A8268}
-	EndGlobalSection
-EndGlobal
