+++ README.md
@@ -3,6 +3,8 @@ Bypass EDR Hooks by patching NT API stub, and resolving SSNs and syscall instruc
 
 ![image](https://raw.githubusercontent.com/illegal-instruction-co/UnhookingPatch/main/assets/view.jpg)
 
+## Building
+Run build.bat if you are building it first time, then you can build it after changes from **VS** command prompt.
 
 ## How do i convert binary to MAC ?
 
@@ -11,6 +13,7 @@ Requirements:
 
 ```
 pip install macaddress
-./bin2mac.py calc.bin
+./tools/bin2mac.py shellcode_file.exe
 ```
-
+Then check
+./tools/shellcode_file.exe.c
